# applications/oxybank/web/openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/trigger/history/{kb_name}": {
      "get": {
        "tags": [
          "trigger"
        ],
        "summary": "Query History By Kb",
        "description": "Query trigger execution history for a knowledge base.\n\nArgs:\n    kb_name: Knowledge base name\n    trigger_id: Optional filter by trigger ID\n    status: Optional filter by execution status (pending/success/failed)\n    start_date: Optional start date filter\n    end_date: Optional end date filter\n    page: Page number (1-based)\n    page_size: Records per page (max 100)\n\nReturns:\n    APIResponse with execution history",
        "operationId": "query_history_by_kb_api_v1_trigger_history__kb_name__get",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          },
          {
            "name": "trigger_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by trigger ID",
              "title": "Trigger Id"
            },
            "description": "Filter by trigger ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by execution status",
              "title": "Status"
            },
            "description": "Filter by execution status"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (YYYY-MM-DD HH:MM:SS)",
              "title": "Start Date"
            },
            "description": "Start date (YYYY-MM-DD HH:MM:SS)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (YYYY-MM-DD HH:MM:SS)",
              "title": "End Date"
            },
            "description": "End date (YYYY-MM-DD HH:MM:SS)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Records per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Records per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trigger/history/{kb_name}/{trigger_id}": {
      "get": {
        "tags": [
          "trigger"
        ],
        "summary": "Query History By Trigger",
        "description": "Query execution history for a specific trigger.\n\nArgs:\n    kb_name: Knowledge base name\n    trigger_id: Trigger ID\n    status: Optional filter by execution status\n    start_date: Optional start date filter\n    end_date: Optional end date filter\n    page: Page number (1-based)\n    page_size: Records per page (max 100)\n\nReturns:\n    APIResponse with execution history",
        "operationId": "query_history_by_trigger_api_v1_trigger_history__kb_name___trigger_id__get",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          },
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trigger Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by execution status",
              "title": "Status"
            },
            "description": "Filter by execution status"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (YYYY-MM-DD HH:MM:SS)",
              "title": "Start Date"
            },
            "description": "Start date (YYYY-MM-DD HH:MM:SS)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (YYYY-MM-DD HH:MM:SS)",
              "title": "End Date"
            },
            "description": "End date (YYYY-MM-DD HH:MM:SS)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Records per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Records per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trigger/stats/{kb_name}": {
      "get": {
        "tags": [
          "trigger"
        ],
        "summary": "Get Trigger Stats",
        "description": "Get execution statistics for a knowledge base.\n\nArgs:\n    kb_name: Knowledge base name\n    trigger_id: Optional filter by trigger ID\n\nReturns:\n    APIResponse with execution statistics",
        "operationId": "get_trigger_stats_api_v1_trigger_stats__kb_name__get",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          },
          {
            "name": "trigger_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by trigger ID",
              "title": "Trigger Id"
            },
            "description": "Filter by trigger ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {

              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}